feat: make the engine part of the setup parameters
We add an extension trait to Engine to specify the group element sizes
per curve used. We implement that for each curve we want to support.

Added support for BLS12-381

This allows moving the type constraint from the constructors in BatchedAccumulator and KeyPair to the params, improving ergonomics, and prepares multi-curve support in the CLI